requirements.txt
Django==1.11.4
pytz==2017.2
psycopg2==2.7.3
Pillow==4.2.1
django-imagekit==4.0.1
olefile==0.44
pilkit==2.0
django-admin-sortable2==0.6.15
CommonMark==0.7.4
django-markwhat==1.5.1
